requirements.txt
pandas==2.1.4
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
tenacity==9.1.2
python-json-logger==4.0.0
black==26.1.0
pylint==4.0.4